✨ feat(audio): increase fft spectrum value resolution from 1 to 2 bytes

+++ src/lib/engine/video/tick_units/audio_spectrum_processor.test.ts
@@ -12,7 +12,7 @@ import { assert, describe, it } from "vitest";
 describe("AudioSpectrumProcessor", () => {
     describe("mappedArray", () => {
         // tests relies on this exact array and the fact that it is sorted ascendingly.
-        const a1 = new Uint8Array([10, 20, 30, 40, 50]);
+        const a1 = new Uint16Array([10, 20, 30, 40, 50]);
 
         it("should map to a smaller length", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
@@ -25,7 +25,7 @@ describe("AudioSpectrumProcessor", () => {
         it("should map to a larger length", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
             const mapped = processor.mappedArray(a1, 10);
-            assert.deepEqual(mapped, new Uint8Array([10, 10, 20, 20, 30, 30, 40, 40, 50, 50]));
+            assert.deepEqual(mapped, new Uint16Array([10, 10, 20, 20, 30, 30, 40, 40, 50, 50]));
         });
         it("should map with min and max", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
@@ -38,17 +38,17 @@ describe("AudioSpectrumProcessor", () => {
         it("should handle edge case of min=max", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
             const mapped = processor.mappedArray(a1, 4, 2, 2);
-            assert.deepEqual(mapped, new Uint8Array([30, 30, 30, 30]));
+            assert.deepEqual(mapped, new Uint16Array([30, 30, 30, 30]));
         });
         it("should handle edge case of new_length=1", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
             const mapped = processor.mappedArray(a1, 1);
-            assert.deepEqual(mapped, new Uint8Array([10]));
+            assert.deepEqual(mapped, new Uint16Array([10]));
         });
         it("should handle edge case of new_length=0", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
             const mapped = processor.mappedArray(a1, 0);
-            assert.deepEqual(mapped, new Uint8Array([]));
+            assert.deepEqual(mapped, new Uint16Array([]));
         });
         it("should map backwards if min>max", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
@@ -73,13 +73,13 @@ describe("AudioSpectrumProcessor", () => {
         it("should throw if array is empty and new_length>0", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
             assert.throws(() => {
-                processor.mappedArray(new Uint8Array([]), 3);
+                processor.mappedArray(new Uint16Array([]), 3);
             });
         });
         it("should return empty array if array is empty and new_length=0", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
-            const mapped = processor.mappedArray(new Uint8Array([]), 0);
-            assert.deepEqual(mapped, new Uint8Array([]));
+            const mapped = processor.mappedArray(new Uint16Array([]), 0);
+            assert.deepEqual(mapped, new Uint16Array([]));
         });
         it("should return same array if new_length equals array length", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
@@ -88,15 +88,15 @@ describe("AudioSpectrumProcessor", () => {
         });
         it("should handle single-element array", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
-            const singleElementArray = new Uint8Array([42]);
+            const singleElementArray = new Uint16Array([42]);
             const mapped = processor.mappedArray(singleElementArray, 5);
-            assert.deepEqual(mapped, new Uint8Array([42, 42, 42, 42, 42]));
+            assert.deepEqual(mapped, new Uint16Array([42, 42, 42, 42, 42]));
         });
         it("should handle single-element array with new_length=1", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
-            const singleElementArray = new Uint8Array([42]);
+            const singleElementArray = new Uint16Array([42]);
             const mapped = processor.mappedArray(singleElementArray, 1);
-            assert.deepEqual(mapped, new Uint8Array([42]));
+            assert.deepEqual(mapped, new Uint16Array([42]));
         });
         it("should handle both min max and smaller new_length", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
@@ -109,7 +109,7 @@ describe("AudioSpectrumProcessor", () => {
         it("should handle both min max and larger new_length", async () => {
             const processor = new (await import("./audio_spectrum_processor")).AudioSpectrumProcessor();
             const mapped = processor.mappedArray(a1, 6, 1, 3);
-            assert.deepEqual(mapped, new Uint8Array([20, 20, 30, 30, 40, 40]));
+            assert.deepEqual(mapped, new Uint16Array([20, 20, 30, 30, 40, 40]));
         });
     });
 });
